requirements.txt
asgiref==3.5.2
PyMySQL
djangorestframework
backports.zoneinfo==0.2.1
distlib==0.3.4
Django==4.0.4
django-js-asset==2.0.0   
django-mptt==0.13.4      
filelock==3.7.0
mysqlclient==2.1.0       
Pillow==9.1.1
platformdirs==2.5.2      
python-decouple==3.6     
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1